FNO (RooIn) working with pretrained model
FNO (RooIn) working with pretrained model form Dave (on Lambda machine). Intermediate commit for archive, clean up,  update FnoHydro ... and make code more elegant (base class ...)

+++ fno_hydro/fno_module/FnoHydro.h
@@ -61,7 +61,7 @@ private:
   torch::Device device;
 
   std::unique_ptr<EOS> fnoEOS;
-  double GetTemperatureFromEos(double ed);
+  float GetTemperatureFromEos(float ed);
 
   // Allows the registration of the module so that it is available to be
   // used by the Jetscape framework.
